% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigmoidcpp}
\alias{sigmoidcpp}
\title{Sigmoid function written with sugar.}
\usage{
sigmoidcpp(x)
}
\arguments{
\item{x}{A double.}
}
\description{
Sigmoid function written with sugar.
}
\author{
David Gerard
}

