% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{varbvsnormupdate_cpp}
\alias{varbvsnormupdate_cpp}
\title{Variational update of normal varbvs.}
\usage{
varbvsnormupdate_cpp(X, sigma, sa, logodds, xy, d, alpha, mu, Xr, i)
}
\arguments{
\item{X}{An n by p matrix of predictors. The rows index the observations and the columns index the variables.}

\item{sigma}{A positive numeric. The residual variance.}

\item{sa}{A positive numeric. The prior variance.}

\item{logodds}{A p-vector that contains the prior log-odds of inclusion for each variable.}

\item{xy}{A p-vector containing \code{t(X) \%*\% y}, where \code{y} is a response vector.}

\item{d}{The diagonal of the Gram matrix, \code{d = diag(t(X) \%*\% X)}}

\item{alpha}{A p-vector of inclusion probabilities for each variable.}

\item{mu}{A p-vector of means for each coefficient.}

\item{Xr}{An n-vector with elements \code{X \%*\% (alpha * mu)}}

\item{i}{A p-vector of integers that give the order of updating the coefficients.}
}
\description{
Variational update of normal varbvs.
}
\author{
David Gerard
}

