\name{varbvs.properties}
\alias{nobs.varbvs}
\title{Accessing Properties of Fitted varbvs Models}
\usage{
  \method{nobs}(object, \ldots)
}
\description{
  All these functions are \code{\link{methods}} for class
  \code{"varbvs"} objects.
}
\details{
  The generic accessor functions \code{nobs},...

  TO DO: Add more details here.
}
\seealso{
  
  \code{\link{nobs}}
}
