\name{varbvsplot}
\alias{varbvsplot}
\title{Summarize variable selection results in a single plot.}
\description{
Generate a single plot that summarizes the results of fitting the
Bayesian variable selection model to the data. When the variables are
genetic markers, the groups are chromosomes, and the posterior
probabilities are plotted on the vertical axis (typically on the
logarithmic scale), the figure resembles a "Manhattan plot" typically
used to summarize the results of a genome-wide association study or
quantitative trait locus (QTL) mapping study.
}
\usage{
varbvsplot (fit, score = NULL, groups = NULL, gap = 0,
            col = "midnightblue", vars = NULL, var.labels = NULL,
            var.col = "magenta", pch = 20, xlab = "", ylab = "",
            ltext.args = "col=\"black\",pos=4,cex=0.5", ...)
}
\arguments{
  \item{fit}{Output of function \code{varbvs}.}

  \item{score}{Value to plot on vertical axis. This must be NULL, or a
  numeric vector with one entry for each variable. Ths default is
  \code{score = NULL}. In this case, he posterior inclusion probability
  for each variable is plotted in the vertical axis, in which this
  probability is averaged over hyperparameter settings, treating
  \code{fit$logw} as (unnormalized) log-marginal probabilities. As
  alternative, set \code{score} to the posterior inclusion probabilities
  that ignore correlations, using \code{varbvsindep}.}

  \item{groups}{Group the variables in the plot according to this
  argument. This must be equal to NULL, or a vector with one entry for
  each variable. If groups = NULL, all variables are assigned to a
  single group. This is useful for grouping the genetic markers by
  chromosome in a genome-wide association study.}

  \item{col}{Argument passed to \code{xyplot} specifying color of points.}

  \item{vars}{Indices of variables to highlight and label.}

  \item{var.labels}{Labels that accompany the highlighted variables only.}

  \item{var.col}{Color of points for highlighted variables only.}

  \item{pch}{Argument passed to \code{xyplot} specifying symbol type.}

  \item{score.line}{Draw a red, dotted horizontal line at this location
  on the vertical axis.}
  
  \item{xlab}{Argument passed to \code{xyplot} specifying horizontal
  axis title.}

  \item{ylab}{Argument passed to \code{xyplot} specifying 
  vertical axis title.}

  \item{ltext.args}{Additional arguments passed to \code{ltext} for
  adding the labels in the plot, specified as a string, which is then
  parsed and evaluated using \code{eval(parse(...))}.}

  \item{...}{Additional arguments passed to \code{xyplot}.}
}
\details{
  Note: varbvsplot uses function \code{xyplot} from the \code{lattice}
  package, and \code{as.layer} from the \code{latticeExtra} package.
}  
\value{
  An object of class \code{"trellis"} generated by functions
  \code{xyplot} and \code{as.layer}.
}
\references{P. Carbonetto and M. Stephens (2012). Scalable variational
inference for Bayesian variable selection in regression, and its
accuracy in genetic association studies. \emph{Bayesian Analysis}
\bold{7}, 73--108.}
\author{Peter Carbonetto \email{peter.carbonetto@gmail.com}}
\seealso{\code{varbvs}, \code{summary.varbvs}, \code{varbvsindep}
}
